+++ Dia/DiaCore/Containers/Graphs/GraphC.inl
@@ -448,7 +448,7 @@ namespace Dia
 				{
 					T possibleUnique = At(i);
 					bool foundInList = false;
-					for (size_t j = 0; j < unique.Size(); j++)
+					for (unsigned int j = 0; j < unique.Size(); j++)
 					{
 						if (possibleUnique == unique[j])
 						{
@@ -472,7 +472,7 @@ namespace Dia
 				{
 					T possibleUnique = At(i);
 					bool foundInList = false;
-					for (size_t j = 0; j < unique.Size(); j++)
+					for (unsigned int j = 0; j < unique.Size(); j++)
 					{
 						if (possibleUnique == unique[j])
 						{
